feat: [GSW-2355] Governance Undelegation/Delegation Modal
