+++ packages/web/src/repositories/governance/mock/get-my-delegates-response.json
@@ -1,30 +1,31 @@
 {
   "delegates": [
     {
-      "address": "0x1234567890abcdef1234567890abcdef12345678",
-      "name": "Validator Alpha",
-      "logoURL": "https://example.com/logos/validator-alpha.png",
+      "address": "g1z5cvp07l80pvsu8rvyyncjrrktpzm8rek4nnnp",
+      "name": "Onbloc Node-1",
+      "logoURL": "/gns.svg",
+      "delegateAmount": "150000.25",
+      "delegatedAt": "2024-12-08T09:45:00Z"
+    },
+    {
+      "address": "g1d8muzprhrzlv0fgj3sxepujgs3j02qerv3rak9",
+      "name": "Validator-Alpha",
+      "logoURL": "/gns.svg",
       "delegateAmount": "50000.00",
       "delegatedAt": "2024-12-15T08:30:00Z"
     },
     {
-      "address": "0x9876543210fedcba9876543210fedcba98765432",
-      "name": "StakePool Beta",
-      "logoURL": "https://example.com/logos/stakepool-beta.png",
+      "address": "g16a7etgm9z2r653ucl36rj0l2yqcxgrz2jyegzx",
+      "name": "Validator-Beta",
+      "logoURL": "/atom.svg",
       "delegateAmount": "25000.50",
       "delegatedAt": "2024-12-10T14:20:00Z"
     },
+
     {
-      "address": "0xabcdef1234567890abcdef1234567890abcdef12",
-      "name": "Node Gamma",
-      "logoURL": "https://example.com/logos/node-gamma.png",
-      "delegateAmount": "15000.25",
-      "delegatedAt": "2024-12-08T09:45:00Z"
-    },
-    {
-      "address": "0x5555666677778888999900001111222233334444",
-      "name": "Secure Validator",
-      "logoURL": "https://example.com/logos/secure-validator.png",
+      "address": "g12d65636a6hyudpyc86q3e7p9ypp646ht4g8nxr",
+      "name": "Validator-1",
+      "logoURL": "/cosmos.svg",
       "delegateAmount": "35000.00",
       "delegatedAt": "2024-12-05T16:10:00Z"
     }
