--- packages/web/src/repositories/governance/model/my-undelegates-info.ts
@@ -1,16 +1,0 @@
-export interface MyUnDelegate {
-  address: string;
-  unDelegateAmount: string;
-  logoURL: string;
-  name: string;
-  unlockTime: string;
-  unDelegatedAt: string;
-}
-
-export interface MyUnDelegatesInfo {
-  delegations: MyUnDelegate[];
-}
-
-export const nullMyUnDelegatesInfo: MyUnDelegatesInfo = {
-  delegations: [],
-};
